requirements.txt
kpn-senml
paho-mqtt
